docs: create documetation for Botzo
This commit creates and moves some of the existing documentation to a
centralized place in `/docs` to which material mkdocs is used to
generate beautiful documentation that can be hosted using github pages
by automating a github action to automatically publish it.

Futhermore, the assets have been moved to `/docs/assets`, which is also
part of #6 idea.