feat: improving client into a proper SDK
This commit brings a change in name to the package. Now it is named
botzo_sdk. This will serve the main entry point to interact with the
microcontroller/hardware and offers a BoztoClient which handles all the
communication behind the scenes with the MCU.

The current state is still very limited and still needs development in
order to obtain a stable version.

For now, supported features are connect and disconnect, sending and
receiving packets -> PING command is working. Additionally, it already
handles when a stream of data is enabled, the case for the IMU. However,
BotzoClient is not yet ready for this.